fix: separate category, change SubNav file directory
