+++ lib/constants/category.ts
@@ -1,0 +1,14 @@
+export const categories = [
+  {
+    displayName: 'New Arrivals',
+    urlName: 'new',
+  },
+  {
+    displayName: 'Outerwear',
+    urlName: 'outerwear',
+  },
+  {
+    displayName: 'Top',
+    urlName: 'top',
+  },
+];
